The total paying time will be displayed in the side menu
When the player clicks the menu button, the clock will be set to paused
and the total game time will not be updated until resume is clicked.

The start/pause button works differently, the time will keep rolling
even when the wave is stopped.

The smallest clock multiplier is 0.2 and the biggest is 3. (too big,
enemies will sprint off the grid)